+++ CHANGELOG.md
@@ -51,7 +51,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fix the output of the open flags for the function `renameat2()`
-- Fix the alteration of the `errno` error in the function `__procfdreadlink()`
+- Fix the alteration of the `errno` error in the internal function
+  `__procfdreadlink()` and the functions `getegid()`, `geteuid()`, `getgid()`
+  and `getuid()`
 - Fix the loading of the user ownership from extended attributes of the
   resolved paths in the function `fchownat()`
 - Fix the missing stripping of the chroot directory in the function